# backend/docs/swagger.yaml
basePath: /v1
definitions:
...
  domain.DailyChallengeData:
    properties:
      available:
        type: boolean
      banner_url:
        type: string
      category:
        type: string
      challenge_id:
        type: string
      content_item_id:
        type: string
      date:
        type: string
      message:
        type: string
      moderated:
        type: boolean
      passage_id:
        type: string
      user_status:
        type: string
    type: object
  domain.DailyHistoryItem:
    properties:
      category:
        type: string
      completed_at:
        type: string
      date:
        type: string
      score:
        type: number
      status:
        type: string
    type: object
...
  domain.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
    type: object
...
  service.ContentItem:
    properties:
      audio_url_uk:
        type: string
      audio_url_us:
        type: string
      difficulty:
        type: integer
      focus_phonemes:
        items:
          type: string
        type: array
      id:
        type: string
      ipa:
        type: string
      text:
        type: string
      topic:
        type: string
      type:
        type: string
    type: object
...
  service.FixGuideItem:
    properties:
      examples: {}
      id:
        type: string
      is_simplified:
        type: boolean
      media_url:
        type: string
      tongue_position_vi:
        type: string
    type: object
...
  service.MinimalPairItem:
    properties:
      audio_a_us:
        type: string
      audio_b_us:
        type: string
      category:
        type: string
      explain_vi:
        type: string
      id:
        type: string
      phoneme_a:
        type: string
      phoneme_b:
        type: string
      word_a:
        type: string
      word_b:
        type: string
    type: object
...
  service.PassageItem:
    properties:
      difficulty:
        type: integer
      id:
        type: string
      sentence_count:
        type: integer
      source:
        type: string
      title:
        type: string
      topic:
        type: string
    type: object
  service.PassageSentenceItem:
    properties:
      id:
        type: string
      ipa:
        type: string
      native_audio_url:
        type: string
      order_index:
        type: integer
      text:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: dev@phonara.app
    name: Phonara Team
  description: |-
    Backend API cho ứng dụng luyện phát âm tiếng Anh Phonara.
    Bao gồm: Authentication, Practice Sessions, Error Profile Engine,
    Content, Shadowing, Minimal Pairs, Subscription, Exam (Speechace), Daily Challenge.
  license:
    name: Proprietary
  title: Phonara API
  version: "1.0"
paths:
...
  /content/fix-guide:
    get:
      description: Trả fix guide tiếng Việt cho phoneme/l1_tag. Fallback về simplified
        guide nếu chưa có rich content (BR-L1-02).
      parameters:
      - description: 'IPA phoneme (vd: /θ/)'
        in: query
        name: phoneme
        type: string
      - description: L1 error tag code
        in: query
        name: l1_tag
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Fix guide
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.FixGuideItem'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Fix guide hướng dẫn sửa lỗi phát âm
      tags:
      - Content
  /content/minimal-pairs:
    get:
      description: Lấy cặp âm tối thiểu, filter theo L1 error tag hoặc category chip
        (S15).
      parameters:
      - description: L1 error tag code (final_consonant_deletion, th_sound…)
        in: query
        name: l1_tag
        type: string
      - description: Category (final_sound, th, l_n, v_w, vowel_length)
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Minimal pairs
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.MinimalPairItem'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Danh sách minimal pairs
      tags:
      - Content
  /content/passages:
    get:
      parameters:
      - description: Source (curated, daily, news, dialogue)
        in: query
        name: source
        type: string
      - description: Difficulty level (1–5)
        in: query
        name: difficulty
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Passages
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.PassageItem'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Danh sách shadowing passages
      tags:
      - Content
  /content/passages/{id}/sentences:
    get:
      parameters:
      - description: Passage UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Sentences
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.PassageSentenceItem'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
        "404":
          description: Passage not found
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Câu trong passage
      tags:
      - Content
  /content/sentences:
    get:
      parameters:
      - description: Topic
        in: query
        name: topic
        type: string
      - description: Goal
        in: query
        name: goal
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Sentences list
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.ContentItem'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Danh sách câu luyện
      tags:
      - Content
  /content/words:
    get:
      description: Lấy danh sách words active, filter theo topic/goal/phoneme (FR-WORD-07).
      parameters:
      - description: Topic (interview, travel…)
        in: query
        name: topic
        type: string
      - description: Goal (communication, interview, ielts_toeic, beginner)
        in: query
        name: goal
        type: string
      - description: 'Focus phoneme (vd: /θ/)'
        in: query
        name: phoneme
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Words list
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.ContentItem'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Danh sách từ vựng
      tags:
      - Content
  /daily/history:
    get:
      description: 'Trả 30 ngày gần nhất: date, category, status (completed/missed),
        score (S27 "Thử thách gần đây").'
      produces:
      - application/json
      responses:
        "200":
          description: Challenge history
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.DailyHistoryItem'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Lịch sử Daily Challenge
      tags:
      - Daily
  /daily/today:
    get:
      description: Trả nội dung challenge ngày hôm nay kèm completion status của user
        (FR-DAILY-01, S27).
      produces:
      - application/json
      responses:
        "200":
          description: Today's challenge
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.DailyChallengeData'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - BearerAuth: []
      summary: Daily Challenge hôm nay
      tags:
      - Daily
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"